# Speed up operations that use the `Coord.cells` method for time coordinates