@bouweandela 2022-09-16
I ran a few more experiments, and this implementation is faster if you generate more than roughly 100 cells.