@bouweandela 2022-09-29
Thanks! I added a what's new entry.